refactor[api-server]: auth 패키지 lottery 패키지에서 분리
