+++ src/main/java/com/watermelon/server/auth/annotations/Uid.java
@@ -1,4 +1,4 @@
-package com.watermelon.server.event.lottery.auth.annotations;
+package com.watermelon.server.auth.annotations;
 
 import java.lang.annotation.ElementType;
 import java.lang.annotation.Retention;
