+++ src/main/java/com/watermelon/server/auth/resolver/UidArgumentResolver.java
@@ -1,6 +1,6 @@
-package com.watermelon.server.event.lottery.auth.resolver;
+package com.watermelon.server.auth.resolver;
 
-import com.watermelon.server.event.lottery.auth.annotations.Uid;
+import com.watermelon.server.auth.annotations.Uid;
 import jakarta.servlet.http.HttpServletRequest;
 import lombok.extern.slf4j.Slf4j;
 import org.springframework.core.MethodParameter;
