+++ src/main/java/com/watermelon/server/common/config/WebConfig.java
@@ -1,12 +1,11 @@
 package com.watermelon.server.common.config;
 
 import com.watermelon.server.admin.interceptor.AdminAuthorizationInterceptor;
-import com.watermelon.server.event.lottery.auth.interceptor.LoginCheckInterceptor;
-import com.watermelon.server.event.lottery.auth.resolver.UidArgumentResolver;
+import com.watermelon.server.auth.interceptor.LoginCheckInterceptor;
+import com.watermelon.server.auth.resolver.UidArgumentResolver;
 import lombok.RequiredArgsConstructor;
 import org.springframework.context.annotation.Configuration;
 import org.springframework.web.method.support.HandlerMethodArgumentResolver;
-import org.springframework.web.servlet.config.annotation.CorsRegistry;
 import org.springframework.web.servlet.config.annotation.InterceptorRegistry;
 import org.springframework.web.servlet.config.annotation.WebMvcConfigurer;
