+++ src/main/java/com/watermelon/server/event/link/controller/LinkController.java
@@ -2,7 +2,7 @@ package com.watermelon.server.event.link.controller;
 
 import com.watermelon.server.event.link.service.LinkService;
 import com.watermelon.server.event.link.utils.LinkUtils;
-import com.watermelon.server.event.lottery.auth.annotations.Uid;
+import com.watermelon.server.auth.annotations.Uid;
 import com.watermelon.server.event.link.dto.MyLinkDto;
 import lombok.RequiredArgsConstructor;
 import org.springframework.http.HttpHeaders;
