+++ src/test/java/com/watermelon/server/APITest.java
@@ -18,8 +18,8 @@ import java.nio.charset.StandardCharsets;
 
 import static com.watermelon.server.Constants.*;
 import static com.watermelon.server.common.constants.PathConstants.*;
-import static com.watermelon.server.event.lottery.auth.service.TestTokenVerifier.TEST_UID;
-import static com.watermelon.server.event.lottery.auth.service.TestTokenVerifier.TEST_VALID_TOKEN;
+import static com.watermelon.server.auth.service.TestTokenVerifier.TEST_UID;
+import static com.watermelon.server.auth.service.TestTokenVerifier.TEST_VALID_TOKEN;
 import static org.springframework.restdocs.mockmvc.RestDocumentationRequestBuilders.*;
 import static org.springframework.test.web.servlet.result.MockMvcResultMatchers.*;
