+++ src/test/java/com/watermelon/server/MockLoginInterceptorConfig.java
@@ -1,14 +1,14 @@
 package com.watermelon.server;
 
-import com.watermelon.server.event.lottery.auth.service.TestTokenVerifier;
-import com.watermelon.server.event.lottery.auth.service.TokenVerifier;
+import com.watermelon.server.auth.service.TestTokenVerifier;
+import com.watermelon.server.auth.service.TokenVerifier;
 import com.watermelon.server.event.link.service.LinkService;
 import com.watermelon.server.event.lottery.service.LotteryService;
 import org.mockito.Mockito;
 import org.springframework.boot.test.context.TestConfiguration;
 import org.springframework.context.annotation.Bean;
 
-import static com.watermelon.server.event.lottery.auth.service.TestTokenVerifier.TEST_UID;
+import static com.watermelon.server.auth.service.TestTokenVerifier.TEST_UID;
 
 @TestConfiguration
 public class MockLoginInterceptorConfig {
