+++ src/test/java/com/watermelon/server/integration/PartsIntegrationTest.java
@@ -1,7 +1,7 @@
 package com.watermelon.server.integration;
 
 import com.watermelon.server.BaseIntegrationTest;
-import com.watermelon.server.event.lottery.auth.service.TestTokenVerifier;
+import com.watermelon.server.auth.service.TestTokenVerifier;
 import com.watermelon.server.event.lottery.domain.LotteryApplier;
 import com.watermelon.server.event.parts.domain.LotteryApplierParts;
 import com.watermelon.server.event.parts.domain.Parts;
@@ -12,7 +12,7 @@ import org.junit.jupiter.api.Test;
 
 import static com.watermelon.server.Constants.*;
 import static com.watermelon.server.common.constants.PathConstants.PARTS;
-import static com.watermelon.server.event.lottery.auth.service.TestTokenVerifier.TEST_UID;
+import static com.watermelon.server.auth.service.TestTokenVerifier.TEST_UID;
 import static org.springframework.test.web.servlet.request.MockMvcRequestBuilders.post;
 import static org.springframework.test.web.servlet.result.MockMvcResultMatchers.status;
