+++ src/test/java/com/watermelon/server/lottery/auth/interceptor/LoginCheckInterceptorTest.java
@@ -1,8 +1,8 @@
 package com.watermelon.server.lottery.auth.interceptor;
 
-import com.watermelon.server.event.lottery.auth.exception.AuthenticationException;
-import com.watermelon.server.event.lottery.auth.interceptor.LoginCheckInterceptor;
-import com.watermelon.server.event.lottery.auth.service.TokenVerifier;
+import com.watermelon.server.auth.exception.AuthenticationException;
+import com.watermelon.server.auth.interceptor.LoginCheckInterceptor;
+import com.watermelon.server.auth.service.TokenVerifier;
 import com.watermelon.server.event.link.service.LinkService;
 import com.watermelon.server.event.lottery.service.LotteryService;
 import jakarta.servlet.http.HttpServletRequest;
@@ -18,7 +18,7 @@ import org.mockito.junit.jupiter.MockitoExtension;
 
 import static com.watermelon.server.Constants.*;
 import static com.watermelon.server.common.constants.HttpConstants.HEADER_LINK_ID;
-import static com.watermelon.server.event.lottery.auth.service.TestTokenVerifier.TEST_UID;
+import static com.watermelon.server.auth.service.TestTokenVerifier.TEST_UID;
 import static org.mockito.Mockito.mock;
 import static org.mockito.Mockito.verify;
