+++ src/test/java/com/watermelon/server/lottery/parts/service/LotteryApplierPartsServiceImplTest.java
@@ -18,7 +18,7 @@ import java.util.List;
 import java.util.Optional;
 
 import static com.watermelon.server.Constants.TEST_PARTS_ID;
-import static com.watermelon.server.event.lottery.auth.service.TestTokenVerifier.TEST_UID;
+import static com.watermelon.server.auth.service.TestTokenVerifier.TEST_UID;
 import static org.assertj.core.api.Assertions.assertThat;
 import static org.mockito.ArgumentMatchers.anyLong;
 import static org.mockito.Mockito.*;
